💄 Website refresh (#6)
* Add .idea to .gitignore file

In this commit, .idea directory has been added to the .gitignore file. This is necessary to prevent IDEA-specific files from being tracked by git.

* Remove 'Learn More' button and down arrow image

The 'Learn More' button and the down arrow image in the index.html file have been commented out. This change will make the elements invisible on the webpage. The update is part of a layout modification, removing non-essential elements for a cleaner interface.

* Update resource link for Clean Architecture learning

The resource link in the FAQs for learning more about Clean Architecture has been updated. It previously pointed to Jason Taylor's GitHub repository, but now it correctly points to SSW's own Clean Architecture repository.

* Update video and link styles on homepage

The homepage has been updated to display a more informative title for the embedded video and the video content itself has been swapped for a more relevant one. Additionally, the styling for some links has been changed to enhance their visibility.

* Comment out "GraphQL" section as Northwind365 does not use GraphQL.

The section about "GraphQL" in the index.html file has been commented out. This change will temporarily remove the section from being visible on the webpage, but it can easily be brought back if needed in the future.

* Added chevron back in which drops the page down to the next section.

* Updated logo.  Added new learn more section.

* Updates from designer feedback

* Review changes

* Fixed CSS

* Fixed css